config files devided
