+++ srcs/main.cpp
@@ -11,6 +11,9 @@
 /* ************************************************************************** */
 
 #include "webserv.h"
+#include "Location.hpp"
+#include "Server.hpp"
+#include "ServerConfig.hpp"
 
 #define PORT 8080
 # define BUF_SIZE 100
@@ -119,156 +122,6 @@ void startServer()
 	close(serv_sock);
 }
 
-class Location
-{
-	private:
-		std::map<std::string, std::string> option;
-
-
-	public:
-		Location() {}
-		Location(Location const &co) {}
-		Location &operator=(Location const &co)
-		{
-			return *this;
-		}
-		~Location() {}
-};
-
-class Server
-{
-	private:
-		std::map<std::string, std::string> option;
-		std::vector<Location> location;
-
-
-	public:
-		//std::string 형태로 인자 받아서 바로 파싱해서 저장하자
-		Server() {}
-		Server(const std::string &str)
-		{
-		}
-		Server(Server const &co) : option(co.option), location(co.location)
-		{}
-		Server &operator=(Server const &co) 
-		{
-			option = co.option;
-			location = co.location;
-			return *this;
-		}
-		~Server() {}
-};
-
-class ServerConfig
-{
-	private:
-		std::map<std::string, std::string> option; // 기본 option 저장
-		std::vector<Server> server;
-
-	public:
-		ServerConfig() {}
-		ServerConfig(ServerConfig const &co) : option(co.option), server(co.server)
-		{}
-		ServerConfig &operator=(ServerConfig const &co)
-		{
-			option = co.option;
-			server = co.server;
-			return *this;
-		}
-		~ServerConfig() {}
-
-		// 
-		int checkLine(std::string line)
-		{
-			int i = 0;
-			int res = 1;
-			while (line[i])
-			{
-				if (ft_isalnum(line[i]))
-					res = 0;
-				i++;
-			}
-			return res;
-		}
-
-		// option 파일의 option 저장
-		void configParse(std::string line)
-		{
-			int index;
-			if ((index = line.find('#')) != -1)
-			{
-				line = line.substr(0, index);
-			}
-			if (checkLine(line))
-				return ;
-			std::pair<std::string, std::string> res;
-			line = line.substr(0, line.find(';'));
-
-			char *test = std::strtok((char *)line.c_str(), " \t");
-			if (test == NULL)
-				return ;
-			res.first = std::string(test);
-
-			test = std::strtok(NULL, " \t");
-			line.clear();
-			while (test != 0)
-			{
-				line.append(test);
-				test = std::strtok(NULL, " \t");
-				if (test != 0)
-					line.append(" ");
-			}
-			res.second = line;
-			if (res.second.empty() || res.first.empty())
-				return ;
-			option[res.first] = res.second; // 덮어쓸 땐 이걸로 바꾸면 됨
-		}
-
-		void saveConfig(int fd)
-		{
-			int parenFlag = 0;
-			char *line;
-			std::string stdline;
-			while (get_next_line(fd, &line))
-			{
-				std::string test(line);
-				if (test.find('{') != -1)
-				{
-					stdline.append(test + '\n');
-					parenFlag++;
-				}
-				else if (test.find('}') != -1)
-				{
-					stdline.append(test + '\n');
-					parenFlag--;
-					if (parenFlag == 0)
-					{
-						server.push_back(Server(stdline));
-						stdline.clear();
-					}
-				}
-				else if (parenFlag > 0)
-				{
-					stdline.append(test + '\n');
-				}
-				else
-				{
-					configParse(test);
-				}
-				// flag == 0 이고 줄에 중괄호가 없으면 config에 값 저장
-				// 줄에 { -> 중괄호 시작이 있으면 flag += 1 && line저장 시작
-				// flag > 0 이고 줄에 중괄호가 없으면 line 저장 후 한번에 하위 저장클래스에 넘겨주기
-				// flag > 0 이고 중괄호 } 가 나오면 flag -= 1 이 때 flag 가 0이되면 이대로 끝내서 하위클래스에 패스
-				free(line);
-			}
-
-			std::map<std::string, std::string>::iterator it;
-
-			for(it = option.begin(); it != option.end(); it++)
-				std::cout << "{" << it->first << "} {" << it->second << "}" << std::endl;
-		}
-
-};
 
 int		main(int argc, char **argv)
 {
